+++ lib/moneta/api/data_mapper.rb
@@ -44,7 +44,6 @@ module Moneta
         properties.each do |property, type|
           if data.has_key?(property)
             value = data[property]
-
             property_value = type.nil? ? value : build_complex_value(type, value)
             instance_variable_set("@#{ property }", property_value)
           end
@@ -57,7 +56,7 @@ module Moneta
         properties.each_with_object({}) do |(property, _), hash|
           value = send(property)
           unless value.nil?
-            hash[ property.to_s.camelize_with_lower ] = to_hash_complex_value(value)
+            hash[ property.to_s.camelize(:lower) ] = to_hash_complex_value(value)
           end
         end
       end