+++ spec/lib/moneta/api/requests/find_bank_accounts_request_spec.rb
@@ -8,13 +8,13 @@ describe Moneta::Api::Requests::FindBankAccountsRequest, vcr: true do
 
   subject { response.bank_account }
 
-  its(:count) { is_expected.to eq 4 }
+  its(:count) { is_expected.to eq 6 }
 
   describe 'first account' do
     let(:bank_account) { response.get_account(1182) }
     subject { bank_account }
 
-    its(:id) { is_expected.to eq '1182' }
+    its(:id) { is_expected.to eq 1182 }
 
     describe '.attribute' do
       subject { bank_account.attribute }
@@ -23,7 +23,7 @@ describe Moneta::Api::Requests::FindBankAccountsRequest, vcr: true do
       its([:account]) { is_expected.to eq(value: '30301810000006000001', approved: false) }
       its([:bik]) { is_expected.to eq(value: '044525225', approved: false) }
       its([:corr_account]) { is_expected.to eq(value: '30101810400000000225', approved: false) }
-      its([:is_international]) { is_expected.to eq(value: false, approved: false) }
+      its([:is_international]) { is_expected.to eq(value: 'false', approved: false) }
     end
   end
 end