+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 2.0.0
+
+- Перевод взаимодействия с Moneta.ru с SOAP на JSON. Это дало прирост в скорости примерно в 3-5 раз.
+  
+  Interaction with MONETA.RU changed from SOAP to JSON.
+  
+- Подключены зависимости Oj и ActiveSupport.
+  
+  Oj and ActiveSupport added as dependencies.
+  
+- Изменились настройки подключения. Доступные опции: `demo_mode`, `logger`, `log_body`, `open_timeout`, `filter`.
+  
+  Connection settings changed. Available options: `demo_mode`, `logger`, `log_body`, `open_timeout`, `filter`.
+
 ## 1.13.0
 
 - Поддержка метода `VerifyPaymentBatchRequest`. Added support for `VerifyPaymentBatchRequest` method. [#50]